create Heads component -refactor UserFave and Dashboard components
